# src/lib/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Expense App API",
    "version": "1.0.0",
    "description": "API documentation for Expense App"
  },
  "servers": [{ "url": "http://localhost:3000/api" }],
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserRegister" }
            }
          }
        },
        "responses": {
          "200": { "description": "User created" },
          "400": { "description": "Username already exists or missing fields" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserLogin" }
            }
          }
        },
        "responses": { "200": { "description": "Logged in, returns JWT cookie" } }
      }
    },
    "/expenses": {
      "get": { "summary": "Get all expenses", "responses": { "200": { "description": "List of expenses" } } },
      "post": {
        "summary": "Create a new expense",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/Expense" } }
          }
        },
        "responses": { "200": { "description": "Expense created" } }
      }
    },
    "/expenses/{id}": {
      "get": { "summary": "Get expense by ID", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Expense details" } } },
      "put": { "summary": "Update expense by ID", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Updated expense" } } },
      "delete": { "summary": "Delete expense by ID", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Deleted" } } }
    }
  },
  "components": {
    "schemas": {
      "UserRegister": { "type": "object", "properties": { "username": { "type": "string" }, "password": { "type": "string" } } },
      "UserLogin": { "type": "object", "properties": { "username": { "type": "string" }, "password": { "type": "string" } } },
      "Expense": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "category": { "type": "string" },
          "amount": { "type": "number" },
          "isRecurring": { "type": "boolean" },
          "taxPercent": { "type": "number" },
          "discount": { "type": "number" }
        }
      }
    }
  }
}